Rename components to better reflect what the map is.
